+++ 1.20.1/Common/src/main/java/fuzs/diagonalblocks/api/v2/impl/DiagonalBlockTypeImpl.java
@@ -9,12 +9,11 @@ import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.properties.Property;
 
 import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
-import java.util.function.IntSupplier;
-import java.util.function.Supplier;
 import java.util.function.UnaryOperator;
 
 public class DiagonalBlockTypeImpl implements DiagonalBlockType {
@@ -25,20 +24,20 @@ public class DiagonalBlockTypeImpl implements DiagonalBlockType {
     private final ResourceLocation name;
     private final Class<? extends Block> targetType;
     private final UnaryOperator<Block> factory;
-    private final IntSupplier blockStateProperties;
+    private final int blockPropertiesCount;
     private final TagKey<Block> blacklistTagKey;
     private final Map<ResourceLocation, UnaryOperator<Block>> factoryOverrides = Maps.newConcurrentMap();
 
-    public DiagonalBlockTypeImpl(String name, Class<? extends Block> targetType, UnaryOperator<Block> factory, Supplier<Block> referenceBlock) {
-        this(name, targetType, factory, () -> referenceBlock.get().getStateDefinition().getProperties().size());
+    public DiagonalBlockTypeImpl(String name, Class<? extends Block> targetType, UnaryOperator<Block> factory, Property<?>... blockProperties) {
+        this(name, targetType, factory, blockProperties.length);
     }
 
-    public DiagonalBlockTypeImpl(String name, Class<? extends Block> targetType, UnaryOperator<Block> factory, IntSupplier blockStateProperties) {
+    public DiagonalBlockTypeImpl(String name, Class<? extends Block> targetType, UnaryOperator<Block> factory, int blockPropertiesCount) {
         name = name.toLowerCase(Locale.ROOT);
         this.name = new ResourceLocation("diagonal" + name, name);
         this.targetType = targetType;
         this.factory = factory;
-        this.blockStateProperties = blockStateProperties;
+        this.blockPropertiesCount = blockPropertiesCount;
         this.blacklistTagKey = TagKey.create(Registries.BLOCK, this.id("non_diagonal_" + name));
     }
 
@@ -63,9 +62,10 @@ public class DiagonalBlockTypeImpl implements DiagonalBlockType {
             // check that block state properties count matches, we might otherwise run into issues with blocks that extend the base class and expect certain properties to be present
             // especially an issue when running BlockBehaviour.Properties::copy with behavior properties that depend on block state properties, such as BlockBehaviour.Properties::lightLevel
             // checking the count is enough, no need to compare actual properties
-            boolean isTarget = this.targetType.isInstance(block) && this.blockStateProperties.getAsInt() == block.getStateDefinition().getProperties().size();
+            boolean isTarget = this.targetType.isInstance(block) && this.blockPropertiesCount == block.getStateDefinition().getProperties().size();
             return isTarget || this.factoryOverrides.containsKey(resourceLocation);
         }
+
         return false;
     }