+++ 1.20.1/Common/src/main/java/fuzs/diagonalblocks/api/v2/impl/DiagonalGlassPaneBlock.java
@@ -95,6 +95,7 @@ public class DiagonalGlassPaneBlock extends IronBarsBlock implements StarCollisi
             }
             stateShapes[i] = voxelCollection.optimize();
         }
+
         return stateShapes;
     }
 }