+++ 1.20.1/Common/src/main/java/fuzs/diagonalblocks/data/client/DynamicModelProvider.java
@@ -33,6 +33,6 @@ public class DynamicModelProvider extends AbstractModelProvider {
 
     @Override
     public void addItemModels(ItemModelBuilder builder) {
-
+        // NO-OP
     }
 }