+++ 1.20.1/CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v8.0.5-1.20.1] - 2024-03-29
+### Changed
+- A few minor improvements towards better mod compatibility:
+  - Prevent invalid model warnings for built-in blacklisted blocks
+  - Disable Better End / Better Nether blocks from generating diagonal variants, instead of only blacklisting those blocks in the corresponding tags
+  - Blacklist blocks from the Domum Ornamentum mod by default
+
 ## [v8.0.4-1.20.1] - 2023-12-08
 ### Fixed
 - Fences from [Immersive Engineering](https://www.curseforge.com/minecraft/mc-mods/immersive-engineering) are now blacklisted by default to prevent issues with multi-blocks not forming