+++ 1.20.4/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v20.4.3-1.20.1] - 2024-03-29
+### Changed
+- A few minor improvements towards better mod compatibility:
+    - Prevent invalid model warnings for built-in blacklisted blocks
+    - Disable Better End / Better Nether blocks from generating diagonal variants, instead of only blacklisting those blocks in the corresponding tags
+    - Blacklist blocks from the Domum Ornamentum mod by default
+
 ## [v20.4.2-1.20.4] - 2024-01-30
 ### Fixed
 - Update to latest version of Puzzles Lib